requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-appconf==1.0.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fire==0.5.0
gunicorn==20.1.0
idna==3.4
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-environ==0.4.54
pytz==2022.7.1
rcssmin==1.1.1
requests==2.28.2
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
termcolor==2.2.0
urllib3==1.26.14
whitenoise==6.3.0
